backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.3
cssselect==1.3.0
defusedxml==0.7.1
distro==1.9.0
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jiter==0.10.0
jmespath==1.0.1
kombu==5.5.4
lxml==5.4.0
openai==1.84.0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
playwright==1.52.0
prompt_toolkit==3.0.51
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyee==13.0.0
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
queuelib==1.8.0
redis==6.2.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.13.1
selenium==4.33.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tldextract==5.3.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.11.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
